Time: 3 ms (60.76%), Space: 5.9 MB (41.61%) - LeetHub

+++ 509-fibonacci-number/509-fibonacci-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int fib(int n) {
         if (n == 0) return 0;
-        long x = 0, y = 1;
+        int x = 0, y = 1;
         for (; n > 1; --n) {
             y += x;
             x = y - x;
